cloud/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
requests==2.31.0
numpngw==0.1.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-multipart==0.0.9
passlib==1.7.4
jinja2==3.1.3
cloud-sql-python-connector[pg8000]==1.10.0
google-cloud-storage==2.17.0
markupsafe==2.1.5